+++ Changelog.md
@@ -6,6 +6,7 @@
 - New: Debug feature to help identify other mods preventing players from viewing a Security Camera
 - New: Configuration option for limiting the amount of chunks that may be forceloaded by frame feeds
 - Change: Reworked the model and texture of the Motion-Activated Light
+- Change: Renamed the Reinforced Chain to Reinforced Iron Chain for consistency with vanilla
 - API: New IOwnable#isOwnedBy overload taking a Player and a boolean for checking ownership respecting the new Incognito Mask
 - API: New method Owner#isDefaultOwner to check if the Owner object has no player data associated with it
 - API: New method IPasscodeConvertible#getRequiredKeyPanels to control how many key panel items are consumed when converting this block